project/pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ecommerce"
version = "0.1.0"
description = "متجر إلكتروني احترافي مبني باستخدام Django"
authors = [
    {name = "Saleh", email = "saleh@example.com"},
]
dependencies = [
    "Django>=4.2.7",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.1",
    "django-environ>=0.11.2",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "django-redis>=5.4.0",
    "django-cacheops>=7.0.2",
    "django-elasticsearch-dsl>=7.3.0",
    "django-allauth>=0.57.0",
    "djangorestframework-simplejwt>=5.3.0",
    "stripe>=7.6.0",
    "paypalrestsdk>=1.13.3",
    "Pillow>=10.1.0",
    "django-imagekit>=4.1.0",
    "django-ckeditor>=6.7.0",
    "django-storages>=1.14.2",
    "boto3>=1.34.0",
    "django-anymail>=10.2",
    "django-debug-toolbar>=4.2.0",
    "django-extensions>=3.2.3",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "celery>=5.3.4",
    "flower>=2.0.1",
    "python-decouple>=3.8",
    "requests>=2.31.0",
    "ruff>=0.1.6",
    "pytest>=7.4.3",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
]
requires-python = ">=3.11"

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "T20", "DJ"]
ignore = ["E501", "B008", "DJ001", "A003"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"*/migrations/*" = ["N806", "N801", "N803"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--cov=apps --cov-report=html --cov-report=term-missing"
